Added unit test for db dir creation.
